requirements.txt
fastapi
uvicorn
pillow
opencv-python
svgwrite
potracer
python-multipart
